requirements.txt
efficientnet-pytorch==0.7.1
opencv-python==4.7.0.68
opencv-python-headless==4.7.0.68
openslide-python==1.2.0
pandas==1.3.5
Pillow==9.3.0
pma-python==2.0.0.167
pretrainedmodels==0.7.4
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
sklearn==0.0.post1
tiatoolbox==1.3.1
timm==0.6.12
torch==1.13.1
torchstain==1.3.0
torchvision==0.14.1
tqdm==4.64.1
xlrd==2.0.1
